refactor(SignInForm): split into smaller components
